requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
mysqlclient>=2.2.0

# Filtering and Search
django-filter>=23.5

# CORS Headers
django-cors-headers>=4.3.0

# Celery for Async Tasks
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=22.0.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
isort>=5.13.0
pylint>=3.0.0

# Environment Variables
python-decouple>=3.8

# Utils
python-dateutil>=2.8.2
pytz>=2023.3

# Monitoring & Logging
django-extensions>=3.2.3

# Production Server
gunicorn>=21.2.0
